server/requirements.txt
flask
sqlalchemy
flask_sqlalchemy
flask-cors
PyJWT